# swagger/swagger.yaml
swagger: "2.0"
info:
  description: "Docs for LioKor Mail API"
  version: "1.0.0"
  title: "LioKor Mail"
  contact:
    email: "korolion31@yandex.ru"
host: "api.mail.liokor.ru"
basePath: "/"
...
schemes:
- "https"
paths:
...
  /email/dialogues:
      get:
        tags:
        - "email"
        summary: "Returns dialogues of an authenticated user"
        description: "Must be authenticated"
        operationId: "getDialogues"
        produces:
        - "application/json"
        parameters:
        - name: "amount"
          in: "query"
          description: "amount of dialogues to receive"
          required: false
          type: "integer"
        - name: "last"
          in: "query"
          description: "return dialogues beginning from that desc"
          required: false
          type: "integer"
        responses:
          "200":
            description: "Returns list of dialogues"
          "401":
            description: "Not authenticated"
  /email/emails:
      get:
        tags:
        - "email"
        summary: "Returns emails of the selected dialogue"
        description: "Must be authenticated"
        operationId: "getEmails"
        produces:
        - "application/json"
        parameters:
        - name: "with"
          in: "query"
          description: "2nd user"
          required: true
          type: "integer"
        - name: "amount"
          in: "query"
          description: "amount of emails to receive"
          required: false
          type: "integer"
        - name: "last"
          in: "query"
          description: "return emails beginning from that desc"
          required: false
          type: "integer"
        responses:
          "200":
            description: "Returns list of emails"
          "401":
            description: "Not authenticated"
  /email:
    post:
      tags:
      - "email"
      summary: "Sends email"
      description: "Must be authenticated"
      operationId: "sendEmail"
      parameters:
      - in: "body"
        name: "body"
        description: "email to send"
        required: true
        schema:
          $ref: "#/definitions/email"
      responses:
        "200":
          description: "Email was sent"
        "401":
          description: "Not authenticated"
definitions:
...
  email:
    type: "object"
    required:
    - "recipient"
    - "subject"
    - "body"
    properties:
      recipient:
        type: "string"
      subject:
        type: "string"
      body:
        type: "string"
...